apps/mill-regression-cli/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mill-regression-cli"
version = "0.1.0"
description = "Regression diff table builder for comparing test run artifacts"
authors = ["vm <vm@qpointz.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<4.0"
pyyaml = "^6.0"
click = "^8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^4.0"
black = "^23.0"
mypy = "^1.0"

[tool.poetry.scripts]
regression-diff = "regression_diff.cli:main"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

